implement "exchange" combat result user interaction
still  needs runtime check on allocation "limit"

incurred a large-ish rewrite of the `game play` scene's root node script, basically aggregating the different pieces of state and ephemeral state into one dictionary. ("ephemeral" here is used to signify that some parts of the state do not exist for the entire lifetime of the scene. In other words, the dictionary's keys are not all always present over an entire player turn).

for #5