switch PriorityQueue base class to RefCounted
this should ensure the queue is freed from memory once the function that
created it returns.

as this is used for movement pathfinding, there feasibly are 62 queues
being made per turn, collectively 15 * 62 = 930 in a single game. It's a
good idea to make sure they aren't cluttering up the users' machine's
memory, especially when the change is so trivial to implement.