cursor no longer internally exits a non-readonly state after a tile is clicked on
There must now be an external trigger that provokes a return to readonly
state.

This simplifies "what is going on" in terms of cursor state when the
game needs the player to click multiple times in succession on the
board. For example, the first deplyment phase requires around 25
successive clicks without ever needing to return/revert back to a
readonly state.